+++ firmware/friend_bot/ota_service.h
@@ -1,11 +1,14 @@
 #pragma once
 
+#include <stddef.h>
+
 class OtaService {
  public:
   void begin();
   void update();
 
  private:
+  bool fetchManifestVersion(char* version, size_t versionSize);
   void tryRemoteUpdate();
 
   char ssid_[33] = {};
@@ -13,5 +16,6 @@ class OtaService {
   bool wifiAttempted_ = false;
   bool wifiConnected_ = false;
   bool otaChecked_ = false;
+  bool manifestChecked_ = false;
   unsigned long lastWifiRetryMs_ = 0;
 };